+++ knowledge.py
